refactor: move quotations mock data to src/action
